requirements.txt
fastapi
uvicorn[standard]
langserve
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
sentence-transformers
python-dotenv